common: move xattr -> os/filestore/os_xattr
Aside from unittest_chain_xattr, these functions are only used by
filestore, so it makes sense to move them out of common.
