+++ src/os/filestore/os_xattr.c
@@ -26,7 +26,7 @@
 #error "Your system is not supported!"
 #endif
 
-#include "common/xattr.h"
+#include "os_xattr.h"
 
 /*
  * Sets extended attribute on a file.
